+++ CHANGELOG.md
@@ -7,7 +7,7 @@ Add changes to the "Unreleased Changes" section. Once you create a version (and 
 to a section with the version name.
 
 ## Unreleased Changes
-
+* Add QR code as a scannable image in Sales Invoice Additional Fields.
 ## 0.13.0
 
 * Add automatic ZATCA CLI setup